fix: address CodeRabbit review feedback
- IsCompoundCommand treats newlines as command separators, so multi-line
  'run: |' scripts reject blind arg appends like any compound command
- machine token moves into a per-user 0700 directory with symlink and
  permission verification before trusting or writing contents
- StreamHandler.Flush drains partial final lines from tracked writers
  after execution, so newline-less output reaches the log before
  attachRunOutcome and --tail read it back
- doctor worktree-mapping check warns when any host (not just the
  alphabetically first) shares the main checkout's remote dir
- log run directories tightened to 0700 to match the 0600 run log
- dead-pid lock tests move behind //go:build !windows with a shared
  deadPid helper; the Windows processAlive stub can't exercise them
- pretty-mode CommandComplete contract doc matches the implementation
- e2e JSON assertions use --machine explicitly; TaskLogPath test uses
  filepath.Join; argcheck tests set USERPROFILE for Windows
